SLT, SLTI, SLTIU, SLTU tested

+++ src/apl115/mips_cpu_instructions.cpp
@@ -225,7 +225,7 @@ mips_error SLTI(uint32_t rs, uint32_t rt, uint16_t immed, mips_cpu_impl *state){
 }
 
 mips_error SLTIU(uint32_t rs, uint32_t rt, uint16_t immed, mips_cpu_impl *state){
-	return mips_cpu_set_register(state, rt, (rs < ((uint32_t)((int32_t)((int16_t)immed)))));
+	return mips_cpu_set_register(state, rt, (rs < (uint32_t)immed));
 }
